requirements.txt
streamlit>=1.29.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
openai>=1.6.1
chromadb>=0.4.22
pypdf>=3.17.4
pdfplumber>=0.10.3
pandas>=2.1.4
openpyxl>=3.1.2
python-dotenv>=1.0.0
tiktoken>=0.5.2
faiss-cpu>=1.8.0
sentence-transformers>=2.2.2
